+++ crates/cast/src/main.rs
@@ -4,7 +4,7 @@ use crate::starknet_commands::{
 };
 use anyhow::Result;
 use camino::Utf8PathBuf;
-use cast::helpers::constants::DEFAULT_ACCOUNTS_FILE;
+use cast::helpers::constants::{DEFAULT_ACCOUNTS_FILE, DEFAULT_MULTICALL_CONTENTS};
 use cast::helpers::scarb_utils::{parse_scarb_config, CastConfig};
 use cast::{
     account_file_exists, get_account, get_block_id, get_chain_id, get_provider,
@@ -30,13 +30,19 @@ struct Cli {
     #[clap(short = 'u', long = "url")]
     rpc_url: Option<String>,
 
-    /// Account name to be used for contract declaration; overrides account from Scarb.toml
+    /// Account to be used for contract declaration;
+    /// When using keystore (`--keystore`), this should be a path to account file    
+    /// When using accounts file, this should be an account name
     #[clap(short = 'a', long)]
     account: Option<String>,
 
     /// Path to the file holding accounts info
     #[clap(short = 'f', long = "accounts-file")]
     accounts_file_path: Option<Utf8PathBuf>,
+
+    /// Path to keystore file; if specified, --account should be a path to starkli JSON account file
+    #[clap(short = 'k', long)]
+    keystore: Option<Utf8PathBuf>,
 
     /// If passed, values will be displayed as integers, otherwise as hexes
     #[clap(short, long)]
@@ -87,29 +93,33 @@ async fn main() -> Result<()> {
 
     match cli.command {
         Commands::Declare(declare) => {
-            account_file_exists(&config.accounts_file)?;
-            let chain_id = get_chain_id(&provider).await?;
-            let mut account =
-                get_account(&config.account, &config.accounts_file, &provider, chain_id)?;
-
+            let account = get_account(
+                &config.account,
+                &config.accounts_file,
+                &provider,
+                &cli.keystore,
+            )
+            .await?;
             let mut result = starknet_commands::declare::declare(
                 &declare.contract,
                 declare.max_fee,
-                &mut account,
+                &account,
                 &cli.path_to_scarb_toml,
                 cli.wait,
             )
             .await;
 
             print_command_result("declare", &mut result, cli.int_format, cli.json)?;
-
             Ok(())
         }
         Commands::Deploy(deploy) => {
-            account_file_exists(&config.accounts_file)?;
-            let chain_id = get_chain_id(&provider).await?;
-            let account = get_account(&config.account, &config.accounts_file, &provider, chain_id)?;
-
+            let account = get_account(
+                &config.account,
+                &config.accounts_file,
+                &provider,
+                &cli.keystore,
+            )
+            .await?;
             let mut result = starknet_commands::deploy::deploy(
                 deploy.class_hash,
                 deploy.constructor_calldata,
@@ -120,8 +130,8 @@ async fn main() -> Result<()> {
                 cli.wait,
             )
             .await;
-            print_command_result("deploy", &mut result, cli.int_format, cli.json)?;
 
+            print_command_result("deploy", &mut result, cli.int_format, cli.json)?;
             Ok(())
         }
         Commands::Call(call) => {
@@ -135,45 +145,58 @@ async fn main() -> Result<()> {
                 block_id.as_ref(),
             )
             .await;
-            print_command_result("call", &mut result, cli.int_format, cli.json)?;
 
+            print_command_result("call", &mut result, cli.int_format, cli.json)?;
             Ok(())
         }
         Commands::Invoke(invoke) => {
-            account_file_exists(&config.accounts_file)?;
-            let chain_id = get_chain_id(&provider).await?;
-            let mut account =
-                get_account(&config.account, &config.accounts_file, &provider, chain_id)?;
+            let account = get_account(
+                &config.account,
+                &config.accounts_file,
+                &provider,
+                &cli.keystore,
+            )
+            .await?;
             let mut result = starknet_commands::invoke::invoke(
                 invoke.contract_address,
                 &invoke.function,
                 invoke.calldata,
                 invoke.max_fee,
-                &mut account,
+                &account,
                 cli.wait,
             )
             .await;
-            print_command_result("invoke", &mut result, cli.int_format, cli.json)?;
 
+            print_command_result("invoke", &mut result, cli.int_format, cli.json)?;
             Ok(())
         }
         Commands::Multicall(multicall) => {
             match &multicall.command {
                 starknet_commands::multicall::Commands::New(new) => {
-                    let result = starknet_commands::multicall::new::new(
-                        new.output_path.clone(),
-                        new.overwrite,
-                    )?;
-                    println!("{result}");
+                    if let Some(output_path) = &new.output_path {
+                        let mut result =
+                            starknet_commands::multicall::new::new(output_path, new.overwrite);
+                        print_command_result(
+                            "multicall new",
+                            &mut result,
+                            cli.int_format,
+                            cli.json,
+                        )?;
+                    } else {
+                        println!("{DEFAULT_MULTICALL_CONTENTS}");
+                    }
                 }
                 starknet_commands::multicall::Commands::Run(run) => {
-                    account_file_exists(&config.accounts_file)?;
-                    let chain_id = get_chain_id(&provider).await?;
-                    let mut account =
-                        get_account(&config.account, &config.accounts_file, &provider, chain_id)?;
+                    let account = get_account(
+                        &config.account,
+                        &config.accounts_file,
+                        &provider,
+                        &cli.keystore,
+                    )
+                    .await?;
                     let mut result = starknet_commands::multicall::run::run(
                         &run.path,
-                        &mut account,
+                        &account,
                         run.max_fee,
                         cli.wait,
                     )