+++ crates/forge/src/lib.rs
@@ -15,6 +15,7 @@ use cairo_lang_utils::ordered_hash_map::OrderedHashMap;
 
 use crate::running::run_from_test_case;
 use crate::scarb::{ForgeConfig, StarknetContractArtifacts};
+pub use crate::test_file_summary::TestFileSummary;
 use test_collector::{collect_tests, LinkedLibrary, TestCase};
 
 pub mod pretty_printing;
@@ -23,6 +24,7 @@ pub mod test_case_summary;
 
 mod cheatcodes_hint_processor;
 mod running;
+mod test_file_summary;
 
 /// Configuration of the test runner
 #[derive(Deserialize, Debug, PartialEq, Default)]
@@ -243,40 +245,6 @@ pub fn run(
 
     pretty_printing::print_test_summary(&summaries);
     Ok(summaries)
-}
-
-/// Summary of the test run in the file
-#[derive(Debug, PartialEq, Clone)]
-pub struct TestFileSummary {
-    /// Summaries of each test case in the file
-    pub test_case_summaries: Vec<TestCaseSummary>,
-    /// Status of the runner after executing tests in the file
-    pub runner_exit_status: RunnerStatus,
-    /// Relative path to the test file
-    pub relative_path: Utf8PathBuf,
-}
-
-impl TestFileSummary {
-    fn count_passed(&self) -> usize {
-        self.test_case_summaries
-            .iter()
-            .filter(|tu| matches!(tu, TestCaseSummary::Passed { .. }))
-            .count()
-    }
-
-    fn count_failed(&self) -> usize {
-        self.test_case_summaries
-            .iter()
-            .filter(|tu| matches!(tu, TestCaseSummary::Failed { .. }))
-            .count()
-    }
-
-    fn count_skipped(&self) -> usize {
-        self.test_case_summaries
-            .iter()
-            .filter(|tu| matches!(tu, TestCaseSummary::Skipped { .. }))
-            .count()
-    }
 }
 
 fn run_tests_from_file(