+++ crates/forge/src/main.rs
@@ -3,17 +3,21 @@ use camino::Utf8PathBuf;
 use clap::Parser;
 use include_dir::{include_dir, Dir};
 use scarb_metadata::MetadataCommand;
+
 use std::path::PathBuf;
 use tempfile::{tempdir, TempDir};
 
-use forge::run;
 use forge::{pretty_printing, RunnerConfig};
+use forge::{run, TestFileSummary};
 
 use forge::scarb::{
     corelib_for_package, dependencies_for_package, get_contracts_map, name_for_package,
     paths_for_package, target_name_for_package, try_get_starknet_artifacts_path,
 };
+use forge::test_case_summary::TestCaseSummary;
 use std::process::{Command, Stdio};
+
+mod init;
 
 static PREDEPLOYED_CONTRACTS: Dir = include_dir!("crates/cheatnet/predeployed-contracts");
 
@@ -25,6 +29,9 @@ struct Args {
     /// Use exact matches for `test_filter`
     #[arg(short, long)]
     exact: bool,
+    /// Create a new forge project with <name> in current directory
+    #[clap(long)]
+    init: Option<String>,
 
     /// Stop test execution after first failed test
     #[arg(short = 'x', long)]
@@ -39,8 +46,20 @@ fn load_predeployed_contracts() -> Result<TempDir> {
     Ok(tmp_dir)
 }
 
-fn main_execution() -> Result<()> {
+fn extract_failed_tests(tests_summaries: Vec<TestFileSummary>) -> Vec<TestCaseSummary> {
+    tests_summaries
+        .into_iter()
+        .flat_map(|test_file_summary| test_file_summary.test_case_summaries)
+        .filter(|test_case_summary| matches!(test_case_summary, TestCaseSummary::Failed { .. }))
+        .collect()
+}
+
+fn main_execution() -> Result<bool> {
     let args = Args::parse();
+    if let Some(project_name) = args.init {
+        init::run(project_name.as_str())?;
+        return Ok(true);
+    }
 
     let predeployed_contracts_dir = load_predeployed_contracts()?;
     let predeployed_contracts_path: PathBuf = predeployed_contracts_dir.path().into();
@@ -60,12 +79,10 @@ fn main_execution() -> Result<()> {
         .output()
         .context("Failed to build contracts with Scarb")?;
     if !build_output.status.success() {
-        bail!(
-            "Scarb build didn't succeed:\n\n{}",
-            String::from_utf8_lossy(&build_output.stdout)
-        )
+        bail!("Scarb build did not succeed")
     }
 
+    let mut all_failed_tests = vec![];
     for package in &scarb_metadata.workspace.members {
         let forge_config = forge::scarb::config_from_scarb_for_package(&scarb_metadata, package)?;
         let (package_path, lib_path) = paths_for_package(&scarb_metadata, package)?;
@@ -89,7 +106,7 @@ fn main_execution() -> Result<()> {
             .transpose()?
             .unwrap_or_default();
 
-        run(
+        let tests_file_summaries = run(
             &package_path,
             &package_name,
             &lib_path,
@@ -99,6 +116,9 @@ fn main_execution() -> Result<()> {
             &contracts,
             &predeployed_contracts,
         )?;
+
+        let mut failed_tests = extract_failed_tests(tests_file_summaries);
+        all_failed_tests.append(&mut failed_tests);
     }
 
     // Explicitly close the temporary directories so we can handle the errors
@@ -109,15 +129,18 @@ fn main_execution() -> Result<()> {
         )
     })?;
 
-    Ok(())
+    pretty_printing::print_failures(&all_failed_tests);
+
+    Ok(all_failed_tests.is_empty())
 }
 
 fn main() {
     match main_execution() {
-        Ok(()) => std::process::exit(0),
+        Ok(true) => std::process::exit(0),
+        Ok(false) => std::process::exit(1),
         Err(error) => {
             pretty_printing::print_error_message(&error);
-            std::process::exit(1);
+            std::process::exit(2);
         }
     };
 }