+++ crates/forge/tests/integration/mod.rs
@@ -14,4 +14,5 @@ mod should_panic;
 mod spoof;
 mod spy_events;
 mod syscalls;
+mod test_state;
 mod warp;