+++ CHANGELOG.md
@@ -12,12 +12,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 
 - `deploy_at` cheatcode
+- printing failures summary at the end of an execution
+
+### Fixed
+
+- non-zero exit code is returned when any tests fail
 
 ### Cast
 
 #### Changed
 
 - fixed misleading error message when there was no scarb in PATH and `--path-to-scarb-toml` was passed
+- modified `multicall new` command output, to be in line with other commands outputs
 
 ## [0.5.0] - 2023-08-30
 
@@ -37,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - rename `parse_txt` trait to `read_txt`
 - support for printing in contracts
 - `spoof` cheatcode
+- snforge command-line flag `--init`
 
 ### Cast